# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  sellerId: number;
  /** @nullable */
  sellerName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  pricePerGram: number;
  weightGrams: number;
  stockGrams: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured: boolean;
  isAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  pricePerGram: number;
  weightGrams: number;
  stockGrams: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  pricePerGram?: number;
  weightGrams?: number;
  stockGrams?: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured?: boolean;
  isAvailable?: boolean;
}

export interface CartItem {
  id: number;
  userId: string;
  productId: number;
  product: Product;
  quantityGrams: number;
  createdAt: string;
}

export interface AddToCartBody {
  productId: number;
  quantityGrams: number;
}

export interface UpdateCartItemBody {
  quantityGrams: number;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantityGrams: number;
  pricePerGram: number;
  subtotal: number;
}

export interface Order {
  id: number;
  buyerUserId: string;
  sellerUserId: string;
  status: string;
  totalAmount: number;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type SellerProfileSellerType =
  (typeof SellerProfileSellerType)[keyof typeof SellerProfileSellerType];

export const SellerProfileSellerType = {
  forager: "forager",
  commercial_supplier: "commercial_supplier",
  restaurant: "restaurant",
  individual: "individual",
} as const;

export interface SellerProfile {
  id: number;
  userId: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  sellerType: SellerProfileSellerType;
  /** @nullable */
  imageUrl?: string | null;
  isVerified: boolean;
  /** @nullable */
  rating?: number | null;
  totalSales: number;
  createdAt: string;
}

export type CreateSellerProfileBodySellerType =
  (typeof CreateSellerProfileBodySellerType)[keyof typeof CreateSellerProfileBodySellerType];

export const CreateSellerProfileBodySellerType = {
  forager: "forager",
  commercial_supplier: "commercial_supplier",
  restaurant: "restaurant",
  individual: "individual",
} as const;

export interface CreateSellerProfileBody {
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  sellerType: CreateSellerProfileBodySellerType;
  /** @nullable */
  imageUrl?: string | null;
}

export type UpdateSellerProfileBodySellerType =
  (typeof UpdateSellerProfileBodySellerType)[keyof typeof UpdateSellerProfileBodySellerType];

export const UpdateSellerProfileBodySellerType = {
  forager: "forager",
  commercial_supplier: "commercial_supplier",
  restaurant: "restaurant",
  individual: "individual",
} as const;

export interface UpdateSellerProfileBody {
  displayName?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  sellerType?: UpdateSellerProfileBodySellerType;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryCount {
  category: string;
  count: number;
  avgPricePerGram: number;
}

export interface DashboardSummary {
  totalOrders: number;
  pendingOrders: number;
  totalSpent: number;
  /** @nullable */
  favoriteCategory?: string | null;
  isSeller: boolean;
}

export interface MonthRevenue {
  month: string;
  revenue: number;
}

export interface SellerStats {
  totalRevenue: number;
  totalOrdersFulfilled: number;
  activeListings: number;
  topProducts: Product[];
  recentOrders: Order[];
  revenueByMonth: MonthRevenue[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  order_placed: "order_placed",
  order_status_update: "order_status_update",
  new_listing: "new_listing",
  product_sold: "product_sold",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  message: string;
  timestamp: string;
  /** @nullable */
  orderId?: number | null;
  /** @nullable */
  productId?: number | null;
}

export type ListProductsParams = {
  category?: string;
  sellerId?: string;
  minPrice?: number;
  maxPrice?: number;
  inStock?: boolean;
  search?: string;
};

export type ListOrdersParams = {
  role?: ListOrdersRole;
  status?: string;
};

export type ListOrdersRole =
  (typeof ListOrdersRole)[keyof typeof ListOrdersRole];

export const ListOrdersRole = {
  buyer: "buyer",
  seller: "seller",
} as const;