# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruffleNation Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "201":
          description: Cart item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear the cart
      responses:
        "204":
          description: Cart cleared
  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ---- ORDERS ----
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (checkout from cart)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ---- SELLERS ----
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Product:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        pricePerGram:
          type: number
        weightGrams:
          type: number
        stockGrams:
          type: number
        origin:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        harvestDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - sellerId
        - name
        - category
        - pricePerGram
        - weightGrams
        - stockGrams
        - isFeatured
        - isAvailable
        - createdAt
        - updatedAt
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantityGrams:
          type: number
        createdAt:
          type: string
      required:
        - id
        - userId
        - productId
        - product
        - quantityGrams
        - createdAt
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantityGrams:
          type: number
      required:
        - productId
        - quantityGrams
    UpdateCartItemBody:
      type: object
      properties:
        quantityGrams:
          type: number
      required:
        - quantityGrams
    Order:
      type: object
      properties:
        id:
          type: integer
        buyerUserId:
          type: string
        sellerUserId:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        shippingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - buyerUserId
        - sellerUserId
        - status
        - totalAmount
        - items
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantityGrams:
          type: number
        pricePerGram:
          type: number
        subtotal:
          type: number
      required:
        - id
        - orderId
        - productId
        - productName
        - quantityGrams
        - pricePerGram
        - subtotal
    CreateOrderBody:
      type: object
      properties:
        shippingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: []
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      required:
        - status
...