# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
}