# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  buyerUserId: string;
  sellerUserId: string;
  status: string;
  totalAmount: number;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}