# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TruffleNation Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  sellerId: number;
  /** @nullable */
  sellerName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  pricePerGram: number;
  weightGrams: number;
  stockGrams: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  season?: string | null;
  /** @nullable */
  harvestDate?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured: boolean;
  isAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}